refactor(settings): extract ProfileAvatarCard widget
Move the profile avatar / display name / upload-remove section out of
settings_screen.dart into a self-contained ProfileAvatarCard widget at
lib/features/settings/widgets/. The new widget owns its own state
(avatar URL, display name, upload-in-progress flag) and the four
profile methods (_fetchProfile, _saveDisplayName, _uploadAvatar,
_removeAvatar). settings_screen.dart now renders it with one line.
No visual or behavioral change.

Closes #53
